Improved JSON encode/decode
Fixed issues with formatting
Made the internals testable.

+++ Sources/OpenAIBits/Types/Penalty.swift
@@ -1,8 +1,10 @@
+import Foundation
+
 /// Represents a value that can be between `0` and `1`.
 public struct Penalty: Equatable {
-  public let value: Double
+  public let value: Decimal
   
-  public init(_ value: Double) {
+  public init(_ value: Decimal) {
     self.value = Self.clamp(value)
   }
 }
@@ -12,21 +14,21 @@ extension Penalty {
   ///
   /// - Parameter value: The value to clamp.
   /// - Returns the clamped value.
-  public static func clamp(_ value: Double) -> Double {
-    return min(2.0, max(-2.0, Double(value)))
+  public static func clamp(_ value: Decimal) -> Decimal {
+    return min(2.0, max(-2.0, value))
   }
 }
 
 extension Penalty: ExpressibleByFloatLiteral {
   public init(floatLiteral value: Double) {
-    self.init(value)
+    self.init(Decimal(value))
   }
 }
 
 extension Penalty: Codable {
   public init(from decoder: Decoder) throws {
     let container = try decoder.singleValueContainer()
-    try self.init(container.decode(Double.self))
+    try self.init(container.decode(Decimal.self))
   }
   
   public func encode(to encoder: Encoder) throws {
